Colocate the "talking to C++" layer
