+++ lib/index.js
@@ -1,5 +1,5 @@
 const {PathWatcherManager} = require('./path-watcher-manager')
-const {configure, status} = require('./native-watcher')
+const {configure, status, DISABLE, STDERR, STDOUT} = require('./binding')
 
 // Extended: Invoke a callback with each filesystem event that occurs beneath a specified path.
 //
@@ -60,4 +60,13 @@ function printWatchers () {
   return PathWatcherManager.instance().print()
 }
 
-module.exports = {watchPath, stopAllWatchers, printWatchers, configure, status}
+module.exports = {
+  watchPath,
+  stopAllWatchers,
+  printWatchers,
+  configure,
+  status,
+  DISABLE,
+  STDERR,
+  STDOUT
+}
