+++ lib/native-watcher.js
@@ -1,9 +1,4 @@
-let watcher = null
-try {
-  watcher = require('../build/Release/watcher.node')
-} catch (err) {
-  watcher = require('../build/Debug/watcher.node')
-}
+const binding = require('./binding')
 const {Emitter, CompositeDisposable, Disposable} = require('event-kit')
 
 const ACTIONS = new Map([
@@ -18,11 +13,6 @@ const ENTRIES = new Map([
   [1, 'directory'],
   [2, 'unknown']
 ])
-
-// Private: Logging mode constants
-const DISABLE = Symbol('disable')
-const STDERR = Symbol('stderr')
-const STDOUT = Symbol('stdout')
 
 // Private: Possible states of a {NativeWatcher}.
 const STOPPED = Symbol('stopped')
@@ -64,7 +54,7 @@ class NativeWatcher {
     this.state = STARTING
 
     this.channel = await new Promise((resolve, reject) => {
-      watcher.watch(this.normalizedPath, this.options, (err, channel) => {
+      binding.watch(this.normalizedPath, this.options, (err, channel) => {
         if (err) {
           reject(err)
           return
@@ -160,7 +150,7 @@ class NativeWatcher {
     this.emitter.emit('will-stop')
 
     await new Promise((resolve, reject) => {
-      watcher.unwatch(this.channel, err => (err ? reject(err) : resolve()))
+      binding.unwatch(this.channel, err => (err ? reject(err) : resolve()))
     })
     this.channel = null
     this.state = STOPPED
@@ -200,60 +190,4 @@ class NativeWatcher {
   }
 }
 
-function logOption (baseName, options, normalized) {
-  const value = options[baseName]
-
-  if (value === undefined) return
-
-  if (value === DISABLE) {
-    normalized[`${baseName}Disable`] = true
-    return
-  }
-
-  if (value === STDERR) {
-    normalized[`${baseName}Stderr`] = true
-    return
-  }
-
-  if (value === STDOUT) {
-    normalized[`${baseName}Stdout`] = true
-    return
-  }
-
-  if (typeof value === 'string' || value instanceof String) {
-    normalized[`${baseName}File`] = value
-    return
-  }
-
-  throw new Error(`option ${baseName} must be DISABLE, STDERR, STDOUT, or a filename`)
-}
-
-module.exports = {
-  configure: function (options) {
-    if (!options) {
-      return Promise.reject(new Error('configure() requires an option object'))
-    }
-
-    const normalized = {}
-
-    logOption('mainLog', options, normalized)
-    logOption('workerLog', options, normalized)
-    logOption('pollingLog', options, normalized)
-
-    if (options.pollingThrottle) normalized.pollingThrottle = options.pollingThrottle
-    if (options.pollingInterval) normalized.pollingInterval = options.pollingInterval
-
-    return new Promise((resolve, reject) => {
-      watcher.configure(normalized, err => (err ? reject(err) : resolve(err)))
-    })
-  },
-
-  status: function () {
-    return watcher.status()
-  },
-
-  NativeWatcher,
-  DISABLE,
-  STDERR,
-  STDOUT
-}
+module.exports = {NativeWatcher}
